requirements (1).txt
absl-py==0.7.1
astor==0.7.1
awscli==1.19.48
botocore==1.20.48
cached-property==1.5.2
certifi==2019.3.9
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
cv2-extras==0.5.4
cv2-plt-imshow==0.0.1
cv2-utils==0.2.0
cv2imageload==1.0.6
cv2module==0.1.2
cv2operator==1.0.1
cv2studio==1.0.2
cv2wrap==1.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dlib-bin==19.21.0
dlib-compiled==19.16.99
docutils==0.15.2
Flask==1.0.2
gast==0.2.2
gdcm==1.1
gevent==1.4.0
gluoncv2==0.0.59
google-pasta==0.2.0
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.19.0
h5py==2.9.0
heartpy==1.2.7
icc-rt==2019.0
idna==2.8
imageai==2.1.6
imageio==2.9.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==3.3.0
imutils==0.5.3
intel-numpy==1.15.1
intel-openmp==2021.1.1
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.10.0
joblib==1.0.0
Keras==2.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.1
locustio==0.11.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mkl==2021.1.1
mkl-fft==1.0.6
mkl-random==1.0.1.1
mock==2.0.0
mrcnn==0.2
msgpack==0.6.1
mxnet==1.5.0
networkx==2.5
numpy==1.16.6
onnx==1.8.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opencv-python-inference-engine==4.0.1.2
openvino==2021.2
openvino-python==2021.1
opt-einsum==3.3.0
os.path2==0.0.4
pandas==1.1.5
pbr==5.1.3
Pillow==7.0.0
progress==1.5
protobuf==3.7.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pycparser==2.20
pycv2==0.0.1
pydicom==2.1.2
pylibjpeg==1.1.1
pylibjpeg-libjpeg==1.1.0
pylibjpeg-openjpeg==1.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
rsa==4.7.2
s3transfer==0.3.6
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.2.1
seaborn==0.11.1
six==1.12.0
sklearn==0.0
tbb==2021.1.1
tbb4py==2021.1.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
test-generator==0.1.1
Theano==1.0.5
threadpoolctl==2.1.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.24.1
validators==0.18.1
Werkzeug==0.15.0
wrapt==1.12.1
zipp==3.4.0
